requirements.txt
Django==4.1.1
djangorestframework==3.13.1
pytz==2022.2.1
sqlparse==0.3.0
dj-database-url==0.5.0
gunicorn==19.9.0
psycopg2-binary==2.9.3